Ensure columns in ConversionSchema are in a fixed order Update TableSchema.MapRow to apply time format if provided
